move funcs to page settings and docstrings
