+++ pages/traffic-cameras.js
@@ -1,4 +1,3 @@
-import { useMemo } from "react";
 import PageHead from "../components/PageHead";
 import MapList from "../components/MapList";
 import PopUpHoverContent from "../components/pages/cameras/PopUpHoverContent";
@@ -6,38 +5,25 @@ import PopUpContent from "../components/pages/cameras/PopUpContent";
 import ListItemContent from "../components/pages/cameras/ListItemContent.js";
 import InfoContent from "../components/pages/cameras/InfoContent";
 import { useSocrata } from "../utils/socrata";
-import { CAMERAS_QUERY, CAMERA_COMM_STATUS_QUERY } from "../utils/queries";
+import {
+  CAMERAS_QUERY,
+  CAMERA_COMM_STATUS_MAX_DATE,
+  CAMERA_COMM_STATUS_QUERY,
+} from "../utils/queries";
 import {
   FILTER_SETTINGS,
   SEARCH_SETTINGS,
   LAYER_STYLES,
+  useCommStatus,
+  useMaxDateQuery,
 } from "../page-settings/traffic-cameras";
 
-const simplifyStatus = (status) => {
-  switch (status) {
-    case "online":
-      return status;
-    default:
-      return "offline";
-  }
-};
-
-const useCommStatus = ({ cameras, statuses }) =>
-  useMemo(() => {
-    if (!cameras || !statuses) {
-      return;
-    }
-    // we're updating cameras in-place, but we want to return a new reference
-    // to control loading / render state
-    cameras.features.forEach((feature) => {
-      const id = parseInt(feature.properties.camera_id);
-      const statusMatch = statuses.find(
-        (status) => parseInt(status.device_id) === id
-      );
-      feature.properties.status = simplifyStatus(statusMatch?.status_desc);
-    });
-    return cameras;
-  }, [cameras, statuses]);
+/**
+ * Dashboard that shows traffic cameras with their communication status.
+ * This page uses two data sources: the camera asset records, as well as the
+ * camera device statuses, which are updated daily. See ../utils/queries for
+ * info about these statuses.
+ */
 
 export default function TrafficCameras() {
   const { data: cameras, error } = useSocrata({ ...CAMERAS_QUERY });
